add rudimentary niml parsing
this is planned to be used with niml output
from 3dinfo (not presently implemented)
It will also parse the niml element extracted
from the afni extension that is often saved in
a nifti header.

More extensive plans in this regard is to define
an xml schema for the niml element to allow
niml validation and automated parsing to the level
of typed arrays. This would also allow recursive
niml elements to be parsed with ease.