RF: Do not explicitly set BUILD_SHARED_LIBS, make it an option
Hardcoding forbids even attempt to build static libraries.
https://cmake.org/cmake/help/v3.3/variable/BUILD_SHARED_LIBS.html
says that typically it is used as an option so user could decide.

FWIW currently the build with OFF (for static libraries) leads to
failure since multiple libraries manage to be included to provide the
same symbols (nifti), so it is something TODO in the future.