make include fixes for osx
GL_TEXTURE_3D_EXT symbol not found without glext.h though
this header wasn't used in the linux build. not sure why. may
have to wrap the include in an if statement

although the import scheme for glu and gl changes on mac when using
the framework opengl implementation in the apple sdk, this is not
true if resorting to using Xquartz windowing.