improve the cmake-build's management of AFNI's python code
The logic for managing python code has been moved to the cmakelists file
in the python_scripts directory

Now by default:

The python executable used for the build is hopefully a little more
intuitive. OSX framework python is used as a last resort.  Preference is
given to python interpreters located in an environment (conda or
virtualenv), indeed the build will fail if an environment is activate
but no python interpeter is found in it. The default minimum version is
3.6 (this can be overridden to use any python 3 provided test execution
  or afnipy installation is not required).

By default the AFNI python package is installed into the current python
interpreter in development mode during the build, and installed into the
interpreter's site-packages during installation.

The build target "pytest" now checks that the python currently found and
used for tests does not differ from the one that was used during the
cmake configuration. If the interpreter does change by accident (due to
an accidental change in PATH/environment, or because the first python
found on the PATH does not satisfy the requirements) it can result in an
irritating state where dependencies are installed into the "wrong"
interpreter. For the brave or insistent this check can be removed by
setting FORCE_CURRENT_PY_INTERP_FOR_TESTS to "ON"

set 3.6 as a minimum supported python version for installing

By default AFNI distributes python code in abin, a directory containing
all binaries and scripts. This change does not effect that.  It is
simply to drop the burden of maintenance for the cmake build system and
the circleci testing infrastructure (for example resolving testing
dependencies pre python 3.6 starts to become very difficult).

only find python interpreter if needed in cmake build

change target name for removing python package

refactor test data fetching code a little