adapt to new permissions management and improve dockerfiles
Introduces some adaptations in the cmake docker file to work with the
non-privileged user that was added.

try to use pip for more python dependencies
install pip to /opt:
this avoids having to chown/chmod more than one directory
Also, if the image is used with singularity, the home directory
would likely override some of the files here

The make build is not so readily adapted because of the difficulties of
having an out of source build.
add option for  build dir removal in make image

correctly raise error for failing pipe in cmake dockerfile