improve circleci image caching and WD management
Pass the dev base image cache version as a parameter to
setup_dev_base_image command so that now all images have the same cache
version that is solely set in the workflow section.

Tidy working directory behavior:
Instead of setting working_directory instead rely on the fact that the
default is always ~/afni. It prevents irritating bugs where the
working_directory is not set correctly (although is unfortunately
implicit rather than explicit).