make it hard to accidentally run as root in container
the approach to users and permissions for the afni containers is quite similar
to that used by the jupyter docker stack. See:
https://jupyter-docker-stacks.readthedocs.io/en/latest/

With this change to the entrypoint, it becomes a lot harder to cause
oneself trouble by running the container as root. Even when setting
--user=root and executing bash, the start.sh script is run which
switches to a non-privileged user.

Permissions issues should be fixed by changing the container user id and by
executing the chmod arguments available (-e CHMOD_HOME etc).

The advantage of this approach is that the container will work with no issues
when run as a singluarity container on a host where the user cannot get root
but still wishes to use all of the afni software in the container.