improve motif find module
I am not sure whether there is a better way to solve this problem
that is described in #131 and #158. It seems that it is triggered
in different ways on different systems. I think it can occur for both
afni and suma/ It manifests in similar errors.

A basic description of the problem is that the gui builds find but a segfault occurs when
at runtime. Xt and Xm are involved.

Other details of the error include:
+ Executing the commnd with LD_PRELOAD set to the motif library (Xm)
usually overcomes the issue.
+ providing the --no-as-needed flag to the linker is typically required
to fix the issue.
+ executing "objdump -p suma" (and on osx I think the equivalent would
be "objdump -macho -dylibs-used ./targets_built/suma" though the error
does not seem to occur on osx) shows Xt before Xm in the dynamic section
+ Moving Xm before Xt in the link order (regardless of subsequent
appearances) is required to fix the issue. This is specifically
referring to the gcc link order. The link order specified in cmake may
not necessarily correspond to this.

In addition to adding --no-as-needed as a public link option of motif,
this commit adds Xt as a dependency to help enforce correct link order.

It is worth noting that things can still be broken by various linking patterns in cmake.
It seems a robust solution is to link against mrix (which links
mri,Xm,Xt,and X11 publically). Additionally build time checks of the gui
binaries is likely a useful strategy in making sure this is a build time
failure in future instead of a runtime failure.