improve ci testing infrastructure
raise NotImplementedError for niml.dset for now

Fix bug, kwargs should not be initialized to a mutable type

globbing for files for the data_paths object is a reasonable way to
describe test data dependencies. This is now supported. If the globbing
fails to find any files an error should be raised