circleci: fix homebrew installation of git
not sure why this is occurring now. git version 2.24 is installed
already for 11.2.1 of catalina. While installing other dependencies by
homebrew git is updated but it then fails because it cannot remove 2.24.
Reverting to last version on 10.14 of osx and preventing an error from
being raised when the removal of git fails during the upgrade cleanup.