Add many improvements to testing infrastructure
add installation help to run_afni_tests.py cli

improve behavior of checking the cmake build directory for old builds
etc.

tidy behavior for dependencies (helpful errors, importing only when
needed)

Reduce false positive rate for gui testing. This was working previously
in a restricted set of contexts. It is hopefully more robust now. Race
conditions are resolved and when GL support with Xvfb is not supported
the GUI test is run on the system's X display server (with a display
lock).

improve robustness of parallel test execution. Race conditions for
graphics and environment variables are largely resolved by using a
display lock and the pytest monkeypatch test fixture to isolate certain
aspects of the environment. Additionally the test infrastructure copies
the environment before running the test command to attempt to isolate
any environment changes triggered by the commands themselves. This is
probably only needed in the case of X calls for which the env variables
are modified prior to the subprocess call.

The different situtations that the test suite might be run in are more
carefully defined/supported/tested. In short, running the tests for a
cmake or a make build when it is either installed or not. Attention is
paid to whether the afnipy python package should be installed in these
scenarios and errors are raised appropriately.

pytest is now run as a module of the correct python executable avoiding
an error where the pytest from the wrong interpretter is called in a
subprocess.

timeout functionality is added for testing. This can be very useful some
times for commands that are known to hang across different executable
versions. In the case of RetroTS.py subprocess.wait has a weird
interaction with the plot showing though... so it won't necessarily
always work. By default, subprocess.run is still used.

1d_tool usage has been replaced with libAfni1d, simplifying import
issues.