executor update
Previously a mixture of virtual machine and a minimal alpine
container with docker was used.

This switch makes the circleci executors all the same. The new executor,
created using .docker/circleci_general_container.dockerfile, is an alpine
image with various tools that are required across all the jobs installed.

Apart from making it easier to handle a single executor (now only apk
for package installation) this new setup will be easier to use with local
execution of the individual jobs. A brief summary is at the top of the
dockerfile.

Using an unprivileged user was explored to no avail. So everything is
run as root.