Finalize parallel test execution behavior and data management
Went back and forth between xdist and pytest parallel trying to get
things to work. The main issues that were occurring were that there were
race conditions for xvfb displays, the physical display (on macos), and
data downloading via datalad. The latter two race conditions were
resolved using a resource lock for the process. This works nicely
except that
a) xdist uses more than one process and it fails.
b) pytest parallel still seems to hang sometimes (and it does not seem
   to be associated with data download being required).

Accordingly, filelocking is now used for physical displays, data
download, and xvfb. Xdist is now the plugin used for parallel execution.

Finished off functionality for source-mode "test-data-volume"

circleci uses a remote docker service that requires the use of a data
volume to mount data. It is not especially useful otherwise. In this
situation the user is root... this generates some trickiness. This new
functionality encapsulates this detail in the run_afni_tests.py tool.

Typically, a source-mode of host will directly mount the local test data
into the container along with the rest of the afni source repository.