Overhaul of gui testing
With timeout/command execution/parallelism nicely implemented at this
point this represent an extensive set of fixes for the gui testing.

There were various issues interacting. At this point it is easier to
decipher an error vs a timeout, vs. a race condition due to parallel
testing. Working with commands that have backgrounded processes etc. now
work a lot better.

xeyes was not installed, installed in dev container now.

Carriage returns are not removed so slashes or semicolons need to be
added to the commands.

Port numbers need to be specified robustly to avoid race conditions
(only really applies to Linux at the moment because suma tests are the
only ones using this at the moment and suma runs on the physical display
server on MacOS because the Xvfb version that comes with XQuartz does
not support iGLX from what I can tell).

fix xvfb messiness:
Currently using a forked version of xvfbwrapper (I wanted to add support
for Mac and parallel execution). Well I mis-used my own fork...

If given the display keyword, the Xvfb instance will only try to acquire
the lock for that display or raise an error.