improve handling of afnipy dependency for tests
Data management should not have afnipy as a hard dependency.

afnipy import is no longer tried ini run_afni_tests.py to keep the logic simpler there.

For the typical install using a flat directory structure with abin afnipy
should not be importable. Otherwise it should. There was some slight flaws
in this implementation that are now fixed:
1) pytest when called as a subprocess works out afnipy based on binaries
   on PATH (the environment propagates to the subprocess now).
2) Inappropriate errors for missing afnipy are no longer raised
3) An error is raised if the directory passed using the --abin flag does
   not have binaries in it.