Enables automatic blackening of commits.
This is performed by running the 'darker' tool as a pre-commit
 operation. In order to activate this behavior, a developer must set git
 config to use hooks from .githooks:

'git config core.hooksPath .githooks'

Define black version explicitly. This will also need to be done in ci.
Changes to this should be carefully considered as the black version you
can alter what is defined as 'correct' style. The implications are not
as severe as for a fully blackened codebase but it will still result in
slightly confusing behavior.