Add --runall option
Combinations, created using the get_params_combinations helper function,
offer a nice additional way of expanding tested parameters of a program
but are not run by default (we don't have all day Daniel). These tests
can be selected using the combinations marker but this ends up filtering
out all unmarked tests (and still won't run combination tests that are
marked with slow or veryslow).

Here comes --runall to the rescue. It can be combined with file based
text selection (but not marker or keyword based selection... that
wouldn't make sense).