allow standard search for python interpreter
The python interpreter is required at development time by the cmake
system. The behavior for searching for the interpreter is defined in
cmake/afni_project_dependencies.cmake. By default python interpreters in
environments are used. This behavior can be overwritten by redefining
the Python_FIND_VIRTUALENV cmake variable. Available behaviors are
described in the cmake docs:
https://cmake.org/cmake/help/latest/module/FindPython.html

For example the configure behavior of cmake could be modified with
-DPython_FIND_VIRTUALENV=STANDARD